==> 1_PlugDataMerge.ipynb <==
import pandas as pd
import numpy as np
import glob

# Define path of the Appliances data of 1 household
commonpath='/Users/Desktop/data-x-data/data/House1/01_plugs_csv/Appliance'
pathlist=[]
for i in range(1,8):
    pathlist.append(commonpath+str(i))
pathlist

# Define default values for missing files
missing = [float('NaN')]*86400

# Generate time range
import datetime
from datetime import datetime
date = list(pd.date_range(start = '6/1/2012', end='1/23/2013', freq = 'D'))
date = [str(i)[:10] for i in date]
date[:5]

# Merge all data of 1 household into a dataframe & print out the days missing data
import os
house1 = pd.DataFrame()
for path in pathlist:
    namelist = []
    path=os.path.join(path)
    filenames = glob.glob(path +'/*.csv')
    filenames.sort()
    print(len(filenames))
    df=[]
        
    for filename in filenames:
        i=filename[-14:-4]
        namelist.append(i)
   
    for d in date:
        if d in namelist:
            filepath = os.path.join(path,d+'.csv')
            #print('data of', i, 'is not missing')
            getdata = pd.read_csv(filepath,names=[d])
            l = getdata[d].tolist()                   
            df = df+l
        else:
            print('data of', d ,'is missing for', path[-10:])
            df = df + missing
    print(len(df))
    
    colname = path[-10:]
    house1[colname] = df

house1.head()

timestamp = list(pd.date_range(start='6/1/2012', end='1/24/2013', freq='s'))
del timestamp[len(timestamp)-1]
house1.index = timestamp

house1.rename(columns={'Appliance1':'Fridge',\
                                'Appliance2':'Dryer',\
                                'Appliance3':'Coffee machine',\
                                'Appliance4':'Kettle',\
                                'Appliance5':'Washing machine',\
                                'Appliance6':'PC',\
                                'Appliance7':'Freezer'}, inplace=True)

house1.tail()

# Filter anomalous data and replace with NaN values
house1.replace(to_replace = -1, value = float('NaN'), inplace = True)

# Export to csv file, Note the filepath
col = house1.columns
house1.to_csv(path_or_buf='~/Desktop/data-x-data/ProcessedData/House1/House1_Appliances.csv', sep=',', na_rep='', float_format=None, columns=col, header=True, index=True)
